#636
* Moved the `cp.apple.finalcutpro.ui` library to `cp.ui`, since it
applies across more than just FCPX.
