+++ src/extensions/cp/apple/finalcutpro/export/ExportDialog.lua
@@ -16,11 +16,11 @@
 local log							= require("hs.logger").new("PrefsDlg")
 local inspect						= require("hs.inspect")
 
-local axutils						= require("cp.apple.finalcutpro.axutils")
+local axutils						= require("cp.ui.axutils")
 local just							= require("cp.just")
 
 local SaveSheet						= require("cp.apple.finalcutpro.export.SaveSheet")
-local WindowWatcher					= require("cp.apple.finalcutpro.ui.WindowWatcher")
+local WindowWatcher					= require("cp.apple.finalcutpro.WindowWatcher")
 
 local id							= require "cp.apple.finalcutpro.ids" "ExportDialog"
