+++ src/extensions/cp/apple/finalcutpro/main/Inspector.lua
@@ -18,7 +18,7 @@ local inspect							= require("hs.inspect")
 
 local just								= require("cp.just")
 local prop								= require("cp.prop")
-local axutils							= require("cp.apple.finalcutpro.axutils")
+local axutils							= require("cp.ui.axutils")
 
 local id								= require("cp.apple.finalcutpro.ids") "Inspector"
