+++ src/extensions/cp/apple/finalcutpro/main/Playhead.lua
@@ -13,7 +13,7 @@
 -- EXTENSIONS:
 --
 --------------------------------------------------------------------------------
-local axutils							= require("cp.apple.finalcutpro.axutils")
+local axutils							= require("cp.ui.axutils")
 local geometry							= require("hs.geometry")
 local prop								= require("cp.prop")
