+++ src/extensions/cp/ui/ScrollArea.lua
@@ -4,7 +4,7 @@
 --------------------------------------------------------------------------------
 --------------------------------------------------------------------------------
 
---- === cp.apple.finalcutpro.ui.ScrollArea ===
+--- === cp.ui.ScrollArea ===
 ---
 --- Scroll Area Module.
 
@@ -13,7 +13,7 @@
 -- EXTENSIONS:
 --
 --------------------------------------------------------------------------------
-local axutils							= require("cp.apple.finalcutpro.axutils")
+local axutils							= require("cp.ui.axutils")
 
 --------------------------------------------------------------------------------
 --
