+++ src/plugins/finalcutpro/browser/addnote.lua
@@ -25,7 +25,7 @@ local mouse								= require("hs.mouse")
 
 local config							= require("cp.config")
 local fcp								= require("cp.apple.finalcutpro")
-local axutils 							= require("cp.apple.finalcutpro.axutils")
+local axutils 							= require("cp.ui.axutils")
 
 --------------------------------------------------------------------------------
 --
